factor streams out
